fix(user): change userId type from Long to Integer to match PostgreSQL SERIAL
Schema Alignment:
- PostgreSQL SERIAL type maps to INTEGER (32-bit), not BIGINT (64-bit)
- Hibernate validation was failing: expected BIGINT but found INTEGER

Entity Layer Changes:
- UserAuthEntity.userId: Long → Integer
- UserProfileEntity.profileId: Long → Integer

Domain Layer Changes:
- UserSnapshot.userId: Long → Integer
- User.userId: Long → Integer
- UserRegisteredEvent.userId: Long → Integer

Application Layer Changes:
- UserResponse.userId: Long → Integer

Root Cause:
- JPA @GeneratedValue(strategy = GenerationType.IDENTITY) with Long expects BIGSERIAL
- Database V1.0__Initial_Schema.sql uses SERIAL (INTEGER)
- Mismatch caused: 'found [serial (Types#INTEGER)], but expecting [bigint (Types#BIGINT)]'

Solution:
- Align Java entity types with existing database schema
- No database migration needed (preserve SERIAL for user_id, profile_id)

Technical Impact:
- Max users: 2,147,483,647 (INTEGER max value) - sufficient for enterprise use
- All mappers automatically handle Integer via type inference
- Zero breaking changes to domain logic

Architecture: Maintains Hexagonal + DDD principles
Issue: Fixes Hibernate schema validation error on startup
Breaking: No

+++ src/main/java/com/spacecodee/securityspacee/user/domain/event/UserRegisteredEvent.java
@@ -10,13 +10,13 @@ import lombok.Getter;
 @Getter
 public final class UserRegisteredEvent {
 
-    private final Long userId;
+    private final Integer userId;
     private final String username;
     private final String email;
     private final UserType userType;
     private final Instant occurredOn;
 
-    public UserRegisteredEvent(Long userId, String username, String email, UserType userType) {
+    public UserRegisteredEvent(Integer userId, String username, String email, UserType userType) {
         this.userId = Objects.requireNonNull(userId);
         this.username = Objects.requireNonNull(username);
         this.email = Objects.requireNonNull(email);
