+++ src/main/java/com/spacecodee/securityspacee/user/infrastructure/persistence/jpa/UserProfileEntity.java
@@ -31,7 +31,7 @@ public class UserProfileEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "profile_id")
-    private Long profileId;
+    private Integer profileId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false, unique = true)
